+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 ## 变更日志
 
 ### 其他
-#### 2025-12-03 `✏️ 3+` `➕ 0+` `🗑️ 0+`
+#### 2025-12-03 `✏️ 4+` `➕ 0+` `🗑️ 0+`
+- 🔧  **chore** `release`: v0.0.8
+  > **🕒  19:57** · `➕1` / `➖1`
+  > `👤 Quiteer`  <p602985@yeah.net> ([`ea40020`](https://github.com/QuiteerJs/web/commit/ea400208a28fe4e0912c192d6b81985a63d963f0))
+  - ✏️ [`packages/vite-plugins/package.json`](https://github.com/QuiteerJs/web/blob/ea400208a28fe4e0912c192d6b81985a63d963f0/packages/vite-plugins/package.json)
 - 🔧  **chore** `release`: v0.0.7
   > **🕒  18:43** · `➕1` / `➖1`
   > `👤 Quiteer`  <p602985@yeah.net> ([`b80e131`](https://github.com/QuiteerJs/web/commit/b80e131486d5ac6e7b447f3a87f561ecb910ae0e))
@@ -333,7 +337,17 @@
   - ✏️ [`pnpm-workspace.yaml`](https://github.com/QuiteerJs/web/blob/3a9e13397fe92cc5856e727ecae4b7dab09ee818/pnpm-workspace.yaml)
 
 ### 修复
-#### 2025-12-03 `✏️ 3+` `➕ 0+` `🗑️ 0+`
+#### 2025-12-03 `✏️ 9+` `➕ 1+` `🗑️ 0+`
+- 🐛  **fix** `release`: 解决第一次运行的时候环境变量未生成的bug
+  > **🕒  18:44** · `➕313` / `➖165`
+  > `👤 Quiteer`  <p602985@yeah.net> ([`57e63eb`](https://github.com/QuiteerJs/web/commit/57e63ebe973dc820687fa6c35bf5cd6c0966a060))
+  - ➕ [`packages/vite-plugins/src/bootstrap-env.ts`](https://github.com/QuiteerJs/web/blob/57e63ebe973dc820687fa6c35bf5cd6c0966a060/packages/vite-plugins/src/bootstrap-env.ts)
+  - ✏️ [`CHANGELOG.md`](https://github.com/QuiteerJs/web/blob/57e63ebe973dc820687fa6c35bf5cd6c0966a060/CHANGELOG.md)
+  - ✏️ [`CHANGELOG_TIMELINE.md`](https://github.com/QuiteerJs/web/blob/57e63ebe973dc820687fa6c35bf5cd6c0966a060/CHANGELOG_TIMELINE.md)
+  - ✏️ [`packages/vite-plugins/src/env-config.ts`](https://github.com/QuiteerJs/web/blob/57e63ebe973dc820687fa6c35bf5cd6c0966a060/packages/vite-plugins/src/env-config.ts)
+  - ✏️ [`packages/vite-plugins/src/index.ts`](https://github.com/QuiteerJs/web/blob/57e63ebe973dc820687fa6c35bf5cd6c0966a060/packages/vite-plugins/src/index.ts)
+  - ✏️ [`playground/vite-plugins-test/env.d.ts`](https://github.com/QuiteerJs/web/blob/57e63ebe973dc820687fa6c35bf5cd6c0966a060/playground/vite-plugins-test/env.d.ts)
+  - ✏️ [`playground/vite-plugins-test/vite.config.ts`](https://github.com/QuiteerJs/web/blob/57e63ebe973dc820687fa6c35bf5cd6c0966a060/playground/vite-plugins-test/vite.config.ts)
 - 🐛  **fix** `release`: env插件分词
   > **🕒  17:13** · `➕33` / `➖6`
   > `👤 Quiteer`  <p602985@yeah.net> ([`3287e48`](https://github.com/QuiteerJs/web/commit/3287e480881ab3d7dd7e6ca8c8e495de2675c67f))